+++ test/nodes/case_age_test.rb
@@ -2,13 +2,14 @@
 
 require 'test_helper'
 
-# case anomaly tests
-class AgeTest < Minitest::Test
+# case age tests
+class CasaAgeTest < Minitest::Test
   def test_should_filter_by_case_birth_date_range
     parser = Canql::Parser.new('all cases born between 10/01/2000 and 10/10/2010')
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2010-10-10'] },
                  parser.meta_data['patient.birthdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_case_exect_birth_date
@@ -16,6 +17,7 @@ class AgeTest < Minitest::Test
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2000-01-10'] },
                  parser.meta_data['patient.birthdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_mothers_birth_date_range
@@ -23,6 +25,7 @@ class AgeTest < Minitest::Test
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2010-10-10'] },
                  parser.meta_data['mother.birthdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_mothers_exect_birth_date
@@ -30,6 +33,7 @@ class AgeTest < Minitest::Test
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2000-01-10'] },
                  parser.meta_data['mother.birthdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_case_death_date_range
@@ -37,6 +41,7 @@ class AgeTest < Minitest::Test
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2010-10-10'] },
                  parser.meta_data['patient.deathdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_case_exect_death_date
@@ -44,6 +49,7 @@ class AgeTest < Minitest::Test
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2000-01-10'] },
                  parser.meta_data['patient.deathdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_mothers_death_date_range
@@ -51,6 +57,7 @@ class AgeTest < Minitest::Test
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2010-10-10'] },
                  parser.meta_data['mother.deathdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_mothers_exect_death_date
@@ -58,6 +65,7 @@ class AgeTest < Minitest::Test
     assert parser.valid?
     assert_equal({ Canql::LIMITS => ['2000-01-10', '2000-01-10'] },
                  parser.meta_data['mother.deathdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_case_combined_birth_and_death_date
@@ -69,6 +77,7 @@ class AgeTest < Minitest::Test
                  parser.meta_data['patient.birthdate'])
     assert_equal({ Canql::LIMITS => ['2015-01-01', '2015-01-01'] },
                  parser.meta_data['patient.deathdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_mother_combined_birth_and_death_date
@@ -80,6 +89,7 @@ class AgeTest < Minitest::Test
                  parser.meta_data['mother.birthdate'])
     assert_equal({ Canql::LIMITS => ['2015-01-01', '2015-01-01'] },
                  parser.meta_data['mother.deathdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 
   def test_should_filter_by_case_birth_date_quarter
@@ -88,5 +98,6 @@ class AgeTest < Minitest::Test
     assert_nil parser.failure_reason
     assert_equal({ Canql::LIMITS => ['2015-04-01', '2015-06-30'] },
                  parser.meta_data['patient.birthdate'])
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
   end
 end