+++ test/nodes/main_case_test.rb
@@ -3,12 +3,13 @@
 require 'test_helper'
 
 # main entry point tests
-class MainTest < Minitest::Test
+class MainCaseTest < Minitest::Test
   def test_should_not_filter_all_tumours
     parser = Canql::Parser.new('all cases')
     assert parser.valid?
     assert_instance_of Hash, parser.meta_data
-    assert parser.meta_data.empty?
+    assert_equal 1, parser.meta_data.count
+    assert_equal base_meta_data, parser.meta_data
     assert_nil parser.failure_reason
   end